+++ internal/transfers/transfers_test.go
@@ -17,12 +17,8 @@ import (
 	"testing"
 	"time"
 
-	moovaccounts "github.com/moov-io/accounts/client"
-	"github.com/moov-io/ach"
 	"github.com/moov-io/base"
-	moovcustomers "github.com/moov-io/customers/client"
 	"github.com/moov-io/paygate/internal/accounts"
-	"github.com/moov-io/paygate/internal/customers"
 	"github.com/moov-io/paygate/internal/database"
 	"github.com/moov-io/paygate/internal/depository"
 	"github.com/moov-io/paygate/internal/events"
@@ -774,146 +770,6 @@ func TestTransfers__deleteUserTransfer(t *testing.T) {
 	}
 }
 
-func TestTransfers__validateUserTransfer(t *testing.T) {
-	db := database.CreateTestSqliteDB(t)
-	defer db.Close()
-
-	repo := &SQLRepo{db.DB, log.NewNopLogger()}
-
-	amt, _ := model.NewAmount("USD", "32.41")
-	userID := id.User(base.ID())
-	req := &transferRequest{
-		Type:                   model.PushTransfer,
-		Amount:                 *amt,
-		Originator:             model.OriginatorID("originator"),
-		OriginatorDepository:   id.Depository("originator"),
-		Receiver:               model.ReceiverID("receiver"),
-		ReceiverDepository:     id.Depository("receiver"),
-		Description:            "money",
-		StandardEntryClassCode: "PPD",
-		fileID:                 "test-file",
-	}
-	transfers, err := repo.createUserTransfers(userID, []*transferRequest{req})
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(transfers) != 1 {
-		t.Errorf("got %d transfers", len(transfers))
-	}
-
-	w := httptest.NewRecorder()
-	r := httptest.NewRequest("POST", fmt.Sprintf("/transfers/%s/failed", transfers[0].ID), nil)
-	r.Header.Set("x-user-id", userID.String())
-
-	xferRouter := CreateTestTransferRouter(nil, nil, nil, nil, repo, achclient.AddValidateRoute)
-	defer xferRouter.close()
-
-	router := mux.NewRouter()
-	xferRouter.RegisterRoutes(router)
-	router.ServeHTTP(w, r)
-	w.Flush()
-
-	if w.Code != http.StatusOK {
-		t.Errorf("got %d", w.Code)
-	}
-
-	// have our repository error and verify we get non-200's
-	mockRepo := &MockRepository{Err: errors.New("bad error")}
-	xferRouter.transferRepo = mockRepo
-
-	w = httptest.NewRecorder()
-	router.ServeHTTP(w, r)
-	w.Flush()
-
-	if w.Code != http.StatusBadRequest {
-		t.Errorf("got %d", w.Code)
-	}
-
-	// no repository error, but pretend the ACH file is invalid
-	mockRepo.Err = nil
-	xferRouter2 := CreateTestTransferRouter(nil, nil, nil, nil, repo, achclient.AddInvalidRoute)
-
-	router = mux.NewRouter()
-	xferRouter2.RegisterRoutes(router)
-	w = httptest.NewRecorder()
-	router.ServeHTTP(w, r)
-	w.Flush()
-
-	if w.Code != http.StatusBadRequest {
-		t.Errorf("got %d", w.Code)
-	}
-}
-
-func TestTransfers__getUserTransferFiles(t *testing.T) {
-	db := database.CreateTestSqliteDB(t)
-	defer db.Close()
-
-	repo := &SQLRepo{db.DB, log.NewNopLogger()}
-
-	amt, _ := model.NewAmount("USD", "32.41")
-	userID := id.User(base.ID())
-	req := &transferRequest{
-		Type:                   model.PushTransfer,
-		Amount:                 *amt,
-		Originator:             model.OriginatorID("originator"),
-		OriginatorDepository:   id.Depository("originator"),
-		Receiver:               model.ReceiverID("receiver"),
-		ReceiverDepository:     id.Depository("receiver"),
-		Description:            "money",
-		StandardEntryClassCode: "PPD",
-		fileID:                 "test-file",
-	}
-	transfers, err := repo.createUserTransfers(userID, []*transferRequest{req})
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(transfers) != 1 {
-		t.Errorf("got %d transfers", len(transfers))
-	}
-
-	w := httptest.NewRecorder()
-	r := httptest.NewRequest("POST", fmt.Sprintf("/transfers/%s/files", transfers[0].ID), nil)
-	r.Header.Set("x-user-id", userID.String())
-
-	xferRouter := CreateTestTransferRouter(nil, nil, nil, nil, repo, achclient.AddGetFileRoute)
-	defer xferRouter.close()
-
-	router := mux.NewRouter()
-	xferRouter.RegisterRoutes(router)
-	router.ServeHTTP(w, r)
-	w.Flush()
-
-	if w.Code != http.StatusOK {
-		t.Errorf("got %d", w.Code)
-	}
-
-	bs, _ := ioutil.ReadAll(w.Body)
-	bs = bytes.TrimSpace(bs)
-
-	// Verify it's an array returned
-	if !bytes.HasPrefix(bs, []byte("[")) || !bytes.HasSuffix(bs, []byte("]")) {
-		t.Fatalf("unknown response: %v", string(bs))
-	}
-
-	// ach.FileFromJSON doesn't handle multiple files, so for now just break up the array
-	file, err := ach.FileFromJSON(bs[1 : len(bs)-1]) // crude strip of [ and ]
-	if err != nil || file == nil {
-		t.Errorf("file=%v err=%v", file, err)
-	}
-
-	// have our repository error and verify we get non-200's
-	mockRepo := &MockRepository{Err: errors.New("bad error")}
-	xferRouter.transferRepo = mockRepo
-
-	w = httptest.NewRecorder()
-	router.ServeHTTP(w, r)
-	w.Flush()
-
-	if w.Code != http.StatusBadRequest {
-		t.Errorf("got %d", w.Code)
-	}
-}
-
 func TestTransfers__writeResponse(t *testing.T) {
 	w := httptest.NewRecorder()
 
@@ -958,181 +814,6 @@ func TestTransfers__writeResponse(t *testing.T) {
 	}
 	if pluralResponse[0].ID == "" {
 		t.Errorf("empty transfer: %#v", pluralResponse[0])
-	}
-}
-
-func TestTransfers__createTransactionLines(t *testing.T) {
-	orig := &moovaccounts.Account{ID: base.ID()}
-	rec := &moovaccounts.Account{ID: base.ID()}
-	amt, _ := model.NewAmount("USD", "12.53")
-
-	lines := createTransactionLines(orig, rec, *amt, model.PushTransfer)
-	if len(lines) != 2 {
-		t.Errorf("got %d lines: %v", len(lines), lines)
-	}
-
-	// First transactionLine
-	if lines[0].AccountID != orig.ID {
-		t.Errorf("lines[0].AccountID=%s", lines[0].AccountID)
-	}
-	if !strings.EqualFold(lines[0].Purpose, "ACHDebit") {
-		t.Errorf("lines[0].Purpose=%s", lines[0].Purpose)
-	}
-	if lines[0].Amount != 1253 {
-		t.Errorf("lines[0].Amount=%d", lines[0].Amount)
-	}
-
-	// Second transactionLine
-	if lines[1].AccountID != rec.ID {
-		t.Errorf("lines[1].AccountID=%s", lines[1].AccountID)
-	}
-	if !strings.EqualFold(lines[1].Purpose, "ACHCredit") {
-		t.Errorf("lines[1].Purpose=%s", lines[1].Purpose)
-	}
-	if lines[1].Amount != 1253 {
-		t.Errorf("lines[1].Amount=%d", lines[1].Amount)
-	}
-
-	// flip the TransferType
-	lines = createTransactionLines(orig, rec, *amt, model.PullTransfer)
-	if !strings.EqualFold(lines[0].Purpose, "ACHCredit") {
-		t.Errorf("lines[0].Purpose=%s", lines[0].Purpose)
-	}
-	if !strings.EqualFold(lines[1].Purpose, "ACHDebit") {
-		t.Errorf("lines[1].Purpose=%s", lines[1].Purpose)
-	}
-}
-
-func TestTransfers__postAccountTransaction(t *testing.T) {
-	transferRepo := &MockRepository{}
-
-	xferRouter := CreateTestTransferRouter(nil, nil, nil, nil, transferRepo)
-	defer xferRouter.close()
-
-	if a, ok := xferRouter.accountsClient.(*accounts.MockClient); ok {
-		a.Accounts = []moovaccounts.Account{
-			{
-				ID: base.ID(), // Just a stub, the fields aren't checked in this test
-			},
-		}
-		a.Transaction = &moovaccounts.Transaction{ID: base.ID()}
-	} else {
-		t.Fatalf("unknown accounts.Client: %T", xferRouter.accountsClient)
-	}
-
-	amt, _ := model.NewAmount("USD", "63.21")
-	origDep := &model.Depository{
-		EncryptedAccountNumber: "214124124",
-		RoutingNumber:          "1215125151",
-		Type:                   model.Checking,
-	}
-	recDep := &model.Depository{
-		EncryptedAccountNumber: "212142",
-		RoutingNumber:          "1215125151",
-		Type:                   model.Savings,
-	}
-
-	userID, requestID := id.User(base.ID()), base.ID()
-	tx, err := xferRouter.postAccountTransaction(userID, origDep, recDep, *amt, model.PullTransfer, requestID)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if tx == nil {
-		t.Errorf("nil accounts.Transaction")
-	}
-}
-
-func TestTransfers__verifyCustomerStatus(t *testing.T) {
-	client := &customers.TestClient{
-		Customer: &moovcustomers.Customer{
-			ID:     base.ID(),
-			Status: "ofac",
-		},
-	}
-	orig := &model.Originator{
-		CustomerID: base.ID(),
-	}
-	rec := &model.Receiver{
-		CustomerID: base.ID(),
-	}
-
-	userID := id.User(base.ID())
-
-	if err := verifyCustomerStatuses(orig, rec, client, base.ID(), userID); err != nil {
-		t.Errorf("unexpected error: %v", err)
-	}
-
-	// set an unacceptable status
-	client.Customer.Status = "reviewrequired"
-	if err := verifyCustomerStatuses(orig, rec, client, base.ID(), userID); err == nil {
-		t.Error("expected error")
-	}
-
-	// set an invalid status
-	client.Customer.Status = "invalid"
-	if err := verifyCustomerStatuses(orig, rec, client, base.ID(), userID); err == nil {
-		t.Error("expected error")
-	}
-
-	// set an error and handle it
-	client.Err = errors.New("bad error")
-	if err := verifyCustomerStatuses(orig, rec, client, base.ID(), userID); err == nil {
-		t.Error("expected error")
-	}
-}
-
-func TestTransfers__verifyDisclaimersAreAccepted(t *testing.T) {
-	orig := &model.Originator{CustomerID: base.ID()}
-	rec := &model.Receiver{CustomerID: base.ID()}
-	requestID, userID := base.ID(), id.User(base.ID())
-
-	client := &customers.TestClient{}
-
-	if err := verifyDisclaimersAreAccepted(orig, rec, client, requestID, userID); err != nil {
-		t.Errorf("unexpected error: %v", err)
-	}
-
-	client.Err = errors.New("bad error")
-	if err := verifyDisclaimersAreAccepted(orig, rec, client, requestID, userID); err == nil {
-		t.Error("expected error")
-	}
-}
-
-func TestTransfers__getUserEvents(t *testing.T) {
-	db := database.CreateTestSqliteDB(t)
-	defer db.Close()
-
-	keeper := secrets.TestStringKeeper(t)
-
-	depRepo := depository.NewDepositoryRepo(log.NewNopLogger(), db.DB, keeper)
-
-	transferID := id.Transfer(base.ID())
-	transferRepo := &MockRepository{
-		Xfer: &model.Transfer{
-			ID: transferID,
-		},
-	}
-
-	eventRepo := &events.TestRepository{
-		Event: &events.Event{
-			ID: events.EventID(base.ID()),
-		},
-	}
-	recRepo := &receivers.MockRepository{}
-	origRepo := &originators.MockRepository{}
-
-	router := CreateTestTransferRouter(depRepo, eventRepo, recRepo, origRepo, transferRepo)
-	defer router.close()
-
-	w := httptest.NewRecorder()
-	r := httptest.NewRequest("GET", fmt.Sprintf("/transfers/%s/events", transferID), nil)
-	r.Header.Set("x-user-id", base.ID())
-
-	router.getUserTransferEvents()(w, r)
-	w.Flush()
-
-	if w.Code != http.StatusOK {
-		t.Errorf("got %d", w.Code)
 	}
 }
 
@@ -1200,24 +881,6 @@ func TestTransfers__HTTPDeleteNoUserID(t *testing.T) {
 
 	w := httptest.NewRecorder()
 	r := httptest.NewRequest("DELETE", "/transfers/foo", nil)
-	router.ServeHTTP(w, r)
-	w.Flush()
-
-	if w.Code != http.StatusForbidden {
-		t.Errorf("got %d", w.Code)
-	}
-}
-
-func TestTransfers__HTTPGetEventsNoUserID(t *testing.T) {
-	xfer := CreateTestTransferRouter(nil, nil, nil, nil, nil)
-	defer xfer.close()
-
-	router := mux.NewRouter()
-
-	xfer.RegisterRoutes(router)
-
-	w := httptest.NewRecorder()
-	r := httptest.NewRequest("GET", "/transfers/foo/events", nil)
 	router.ServeHTTP(w, r)
 	w.Flush()
